Allowing the client to specify `activity_date_field`
Previously the date on which the activity was completed was assumed to
be given by the `created_at` field on the `activity_table_name`, i.e.
`user_activity_markers` for us.
This change means we can now configure the analysis to take the custom
field of `user_activity_markers.on_date` to represent when the activity
was actually carried out.  This is more accurate for our considerations.