Updating query to filter results using :activity_date_field
I had previously introduced the :activity_date_field as a way for the
client to customize the field on the :activity_class (e.g.
UserActiveMarker) which should define when the user was active. This was
previously hardcoded as :created_at.

I have forgotten to update the filter on the query so that it users this
new customizable field name, :activity_date_field